reorganize headers, add port support
